Moved tests out of source directory
