+++ conftest.py
@@ -27,8 +27,8 @@ def pytest_configure(config):
     config.data_folder_path = Path("adopt_net0/test/test_data")
     config.result_folder_path = Path("adopt_net0/test/test_results")
     config.case_study_folder_path = Path("adopt_net0/test/test_case")
-    config.technology_data_folder_path = Path("adopt_net0/test/technology_data")
-    config.network_data_folder_path = Path("adopt_net0/test/network_data")
+    config.technology_data_folder_path = Path("tests/technology_data")
+    config.network_data_folder_path = Path("tests/network_data")
     config.root_folder_path = Path(".")
     config.solver = "glpk"
