+++ test_cases/household_minimal/period1/node_data/myHouse/technology_data/Photovoltaic.json
@@ -1,0 +1,32 @@
+{
+  "tec_type": "RES",
+  "size_min": 0,
+  "size_max": 1000,
+  "size_is_int": 0,
+  "decommission": 0,
+  "Economics": {
+    "CAPEX_model": 1,
+    "comment": "CAPEX in EUR/MW, OPEX_variable in EUR/MWh total output, OPEX_fixed in % of CAPEX_annual",
+    "unit_CAPEX": 119362.879,
+    "fix_CAPEX": 0,
+    "OPEX_variable": 16.937521,
+    "OPEX_fixed": 0,
+    "discount_rate": 0.1,
+    "lifetime": 20,
+    "decommission_cost": 0
+  },
+  "Performance": {
+    "comment": "contains fitting data on unit of input, technology types and input/output carriers",
+    "output_carrier": [
+      "electricity"
+    ],
+    "curtailment": 1,
+    "emission_factor": 0
+  },
+  "Units": {
+    "size": "MW",
+    "output_carrier": {
+      "electricity": "MW"
+    }
+  }
+}
