+++ tests/case_study_full_pipeline/ConfigModel.json
