+++ tests/case_study_full_pipeline/NodeLocations.csv
