+++ tests/case_study_full_pipeline/period1/network_topology/existing/size.csv
