+++ tests/case_study_full_pipeline/period1/network_topology/new/electricitySimple/size_max_arcs.csv
