+++ tests/case_study_full_pipeline/period1/node_data/node1/carrier_data/EnergybalanceOptions.json
