+++ tests/case_study_full_pipeline/period1/node_data/node1/technology_data/TestTec_WindTurbine.json
