+++ tests/case_study_full_pipeline/period1/node_data/node2/ClimateData.csv
