+++ tests/case_study_full_pipeline/period1/node_data/node2/technology_data/TestTec_BoilerEl.json
