+++ tests/technology_data/TestTec_GasTurbine_NG_10.json
