+++ tests/technology_data/TestTec_ResPhotovoltaic.json
