+++ tests/test_balances.py
@@ -5,7 +5,7 @@ from pyomo.environ import (
     SolverFactory,
 )
 
-from adopt_net0.test.utilities import make_data_handle
+from tests.utilities import make_data_handle
 from adopt_net0.energyhub import EnergyHub
 from adopt_net0.model_construction.construct_balances import (
     construct_global_balance,
