+++ tests/test_preprocessing.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 import adopt_net0.data_preprocessing as dp
 from adopt_net0.data_management.utilities import check_input_data_consistency
-from adopt_net0.test.utilities import (
+from tests.utilities import (
     select_random_list_from_list,
     load_json,
     save_json,
